four_qubits
A four-qubit system with all-to-all transverse coupling generates a hybridized Hilbert space leading to ZZ splits due to the presence of the higher levels in a qubit (typically transmon). This code estimates the effective ZZ shifts of all four qubits in presence of other qubits. While this theory is general, we have developed this to analyze the multi-qubit processor coupled with a ring bus resonator.